Shared JSON schema utils
